Adds a patch for the FFmpeg trunk to enable ALS demuxing.
This patch has been written by Jai Menon but has not passed review yet.
Refer to thread "[PATCH] als in mp4" for details.

This patch is not meant to be applied as-is in the end but enables
me to continue work on the ALS decoder for the time being. If necessary,
I will continue to work on this patch to pass the review process
if it has not been done as soon as the ALS decoder is to be applied
to the codebase.